slug added

+++ src/main/java/com/nowstartjava/tutorials/service/CategoryService.java
@@ -5,9 +5,12 @@ import java.util.List;
 import com.nowstartjava.tutorials.model.Category;
 
 public interface CategoryService {
-	public Category findOne(long itemId);
+	public Category findOne(int itemId);
+
 	public List<Category> findAll();
+
 	public void save(Category item);
+
 	public void delete(Category category);
 
 }
